+++ Source/APU/Triangle.cpp
@@ -28,7 +28,7 @@ const uint8_t CTriangle::TRIANGLE_WAVE[] = {
 	0x0F, 0x0E, 0x0D, 0x0C, 0x0B, 0x0A, 0x09, 0x08, 0x07, 0x06, 0x05, 0x04, 0x03, 0x02, 0x01, 0x00,
 };
 
-CTriangle::CTriangle(CMixer &Mixer, std::size_t nInstance) :
+CTriangle::CTriangle(CMixer &Mixer, std::uint8_t nInstance) :
 	C2A03Chan(Mixer, {nInstance, sound_chip_t::APU, value_cast(apu_subindex_t::triangle)})		// // //
 {
 	m_iStepGen = 0;
