+++ Source/SoundChipType.h
@@ -41,6 +41,6 @@ public:
 	virtual std::string_view GetChannelShortName(std::size_t subindex) const = 0;
 	virtual std::string_view GetChannelFullName(std::size_t subindex) const = 0;
 
-	virtual std::unique_ptr<CSoundChip> MakeSoundDriver(CMixer &mixer, std::size_t nInstance) const = 0;
-	virtual std::unique_ptr<CChipHandler> MakeChipHandler(std::size_t nInstance) const = 0;
+	virtual std::unique_ptr<CSoundChip> MakeSoundDriver(CMixer &mixer, std::uint8_t nInstance) const = 0;
+	virtual std::unique_ptr<CChipHandler> MakeChipHandler(std::uint8_t nInstance) const = 0;
 };
